Add sequence panel
