+++ packages/core/BaseFeatureWidget/BaseFeatureDetail.tsx
@@ -1,5 +1,5 @@
-/* eslint-disable @typescript-eslint/no-explicit-any,react/prop-types,no-nested-ternary */
-import React, { useEffect, useState } from 'react'
+/* eslint-disable @typescript-eslint/no-explicit-any,react/prop-types */
+import React from 'react'
 import ErrorBoundary from 'react-error-boundary'
 import {
   Accordion,
@@ -8,8 +8,6 @@ import {
   Typography,
   Divider,
   Tooltip,
-  Select,
-  MenuItem,
 } from '@material-ui/core'
 import ExpandMore from '@material-ui/icons/ExpandMore'
 import { makeStyles } from '@material-ui/core/styles'
@@ -17,18 +15,13 @@ import { DataGrid } from '@material-ui/data-grid'
 import { observer } from 'mobx-react'
 import clsx from 'clsx'
 import isObject from 'is-object'
+import { IAnyStateTreeNode } from 'mobx-state-tree'
 import { getConf } from '../configuration'
-import {
-  measureText,
-  getSession,
-  defaultCodonTable,
-  generateCodonTable,
-  revcom,
-} from '../util'
-import { Feature } from '../util/simpleFeature'
+import { measureText, getSession } from '../util'
 import SanitizedHTML from '../ui/SanitizedHTML'
-
-type Feat = { start: number; end: number; type: string }
+import SequenceFeatureDetails from './SequenceFeatureDetails'
+import { BaseCardProps, BaseProps } from './types'
+import { SimpleFeatureSerialized } from '../util/simpleFeature'
 
 const globalOmit = [
   'name',
@@ -97,12 +90,6 @@ export const useStyles = makeStyles(theme => ({
     border: '1px solid #444',
   },
 }))
-
-interface BaseCardProps {
-  title?: string
-  defaultExpanded?: boolean
-  children?: React.ReactNode
-}
 
 export function BaseCard({
   children,
@@ -209,378 +196,6 @@ const ArrayValue = ({
           </div>
         ))
       )}
-    </div>
-  )
-}
-
-interface BaseProps extends BaseCardProps {
-  feature: any
-  descriptions?: Record<string, React.ReactNode>
-  model?: any
-}
-
-function stitch(subfeats: Feat[], sequence: string) {
-  return subfeats.map(sub => sequence.slice(sub.start, sub.end)).join('')
-}
-
-// filter if they have the same ID
-function filterId(feat: Feat) {
-  return `${feat.start}-${feat.end}`
-}
-
-// filters if successive elements share same start/end
-function dedupe(list: Feat[]) {
-  return list.filter(
-    (item, pos, ary) => !pos || filterId(item) !== filterId(ary[pos - 1]),
-  )
-}
-
-function revlist(list: Feat[], seqlen: number) {
-  return list
-    .map(sub => ({
-      ...sub,
-      start: seqlen - sub.end,
-      end: seqlen - sub.start,
-    }))
-    .sort((a, b) => a.start - b.start)
-}
-
-const cdsColor = 'rgba(150,150,0,0.3)'
-const utrColor = 'rgba(0,150,150,0.3)'
-const proteinColor = 'rgba(150,0,150,0.3)'
-const intronColor = 'rgba(200,200,200,0.3)'
-const updownstreamColor = 'rgba(120,0,0,0.3)'
-
-function GeneCDS(props: { cds: Feat[]; sequence: string }) {
-  const { cds, sequence } = props
-
-  return (
-    <div
-      style={{
-        display: 'inline',
-        backgroundColor: cdsColor,
-      }}
-    >
-      {stitch(cds, sequence)}
-    </div>
-  )
-}
-
-function GeneProtein(props: {
-  cds: Feat[]
-  sequence: string
-  codonTable: any
-}) {
-  const { cds, sequence, codonTable } = props
-  const str = stitch(cds, sequence)
-  let protein = ''
-  for (let i = 0; i < str.length; i += 3) {
-    // use & symbol for undefined codon, or partial slice
-    protein += codonTable[str.slice(i, i + 3)] || '&'
-  }
-  return (
-    <div
-      style={{
-        display: 'inline',
-        backgroundColor: proteinColor,
-      }}
-    >
-      {protein}
-    </div>
-  )
-}
-
-function GenecDNA(props: {
-  utr: Feat[]
-  cds: Feat[]
-  exons: Feat[]
-  sequence: string
-  upstream?: string
-  downstream?: string
-  includeIntrons?: boolean
-  collapseIntron?: boolean
-}) {
-  const {
-    utr,
-    cds,
-    exons,
-    sequence,
-    upstream,
-    downstream,
-    includeIntrons,
-    collapseIntron,
-  } = props
-  const chunks = cds.length
-    ? [...cds, ...utr].sort((a, b) => a.start - b.start)
-    : exons
-  return (
-    <>
-      {upstream ? (
-        <div
-          style={{
-            display: 'inline',
-            backgroundColor: updownstreamColor,
-          }}
-        >
-          {upstream}
-        </div>
-      ) : null}
-
-      <>
-        {chunks.map((chunk, index) => {
-          const intron = sequence.slice(chunk.end, chunks[index + 1]?.start)
-          return (
-            <React.Fragment key={JSON.stringify(chunk)}>
-              <div
-                style={{
-                  display: 'inline',
-                  backgroundColor: chunk.type === 'CDS' ? cdsColor : utrColor,
-                }}
-              >
-                {sequence.slice(chunk.start, chunk.end)}
-              </div>
-              {includeIntrons ? (
-                <div
-                  style={{
-                    display: 'inline',
-                    backgroundColor: intronColor,
-                  }}
-                >
-                  {collapseIntron && intron.length > 20
-                    ? `${intron.slice(0, 10)}...${intron.slice(-10)}`
-                    : intron}
-                </div>
-              ) : null}
-            </React.Fragment>
-          )
-        })}
-      </>
-
-      {downstream ? (
-        <div
-          style={{
-            display: 'inline',
-            backgroundColor: updownstreamColor,
-          }}
-        >
-          {downstream}
-        </div>
-      ) : null}
-    </>
-  )
-}
-
-// calculates UTRs using impliedUTRs logic
-function calculateUTRs(cds: Feat[], exons: Feat[]) {
-  const firstCds = cds[0]
-  const lastCds = cds[cds.length - 1]
-  const firstCdsIdx = exons.findIndex(
-    exon => exon.end >= firstCds.start && exon.start <= firstCds.start,
-  )
-  const lastCdsIdx = exons.findIndex(
-    exon => exon.end >= lastCds.end && exon.start <= lastCds.end,
-  )
-  const lastCdsExon = exons[lastCdsIdx]
-  const firstCdsExon = exons[firstCdsIdx]
-
-  const fiveUTRs = [
-    ...exons.slice(0, firstCdsIdx),
-    { start: firstCdsExon.start, end: firstCds.start },
-  ].map(elt => ({ ...elt, type: 'five_prime_UTR' }))
-
-  const threeUTRs = [
-    { start: lastCds.end, end: lastCdsExon.end },
-    ...exons.slice(lastCdsIdx),
-  ].map(elt => ({ ...elt, type: 'three_prime_UTR' }))
-
-  return [...fiveUTRs, ...threeUTRs]
-}
-
-function SequencePanel(props: {
-  sequence: { seq: string; upstream: string; downstream: string }
-  feature: Feat & { strand: number; subfeatures: Feat[] }
-  mode: string
-}) {
-  const { feature, mode } = props
-  let {
-    sequence: { seq: sequence, upstream, downstream },
-  } = props
-  const { subfeatures } = feature
-  const codonTable = generateCodonTable(defaultCodonTable)
-
-  const children = subfeatures
-    .sort((a, b) => a.start - b.start)
-    .map(sub => {
-      return {
-        ...sub,
-        start: sub.start - feature.start,
-        end: sub.end - feature.start,
-      }
-    })
-
-  // we filter duplicate entries in cds and exon lists duplicate entries may
-  // be rare but was seen in Gencode v36 track NCList, likely a bug on GFF3
-  // or probably worth ignoring here (produces broken protein translations if
-  // included)
-  //
-  // position 1:224,800,006..225,203,064 gene ENSG00000185842.15 first
-  // transcript ENST00000445597.6
-  //
-  // http://localhost:3000/?config=test_data%2Fconfig.json&session=share-FUl7G1isvF&password=HXh5Y
-
-  let cds = dedupe(children.filter(sub => sub.type === 'CDS'))
-  let utr = dedupe(children.filter(sub => sub.type.match(/utr/i)))
-  let exons = dedupe(children.filter(sub => sub.type === 'exon'))
-
-  if (!utr.length && cds.length) {
-    utr = calculateUTRs(cds, exons)
-  }
-
-  if (feature.strand === -1) {
-    // doing this in a single assignment is needed because downstream and
-    // upstream are swapped so this avoids a temp variable
-    ;[sequence, upstream, downstream] = [
-      revcom(sequence),
-      revcom(downstream),
-      revcom(upstream),
-    ]
-    cds = revlist(cds, sequence.length)
-    exons = revlist(exons, sequence.length)
-    utr = revlist(utr, sequence.length)
-  }
-
-  if (mode === 'cds') {
-    return <GeneCDS cds={cds} sequence={sequence} />
-  }
-  if (mode === 'cdna') {
-    // utr's were supplied, no inference needed
-    return <GenecDNA exons={exons} cds={cds} utr={utr} sequence={sequence} />
-  }
-  if (mode === 'protein') {
-    return <GeneProtein cds={cds} codonTable={codonTable} sequence={sequence} />
-  }
-
-  if (mode === 'gene') {
-    return (
-      <GenecDNA
-        exons={exons}
-        cds={cds}
-        utr={utr}
-        sequence={sequence}
-        includeIntrons
-      />
-    )
-  }
-
-  if (mode === 'gene_collapsed_intron') {
-    return (
-      <GenecDNA
-        exons={exons}
-        cds={cds}
-        sequence={sequence}
-        utr={utr}
-        includeIntrons
-        collapseIntron
-      />
-    )
-  }
-
-  if (mode === 'gene_updownstream') {
-    return (
-      <GenecDNA
-        exons={exons}
-        cds={cds}
-        sequence={sequence}
-        utr={utr}
-        upstream={upstream}
-        downstream={downstream}
-        includeIntrons
-      />
-    )
-  }
-  return <div>Unknown type</div>
-}
-
-// display the stitched-together sequence of a gene's CDS, cDNA, or protein
-// sequence. this is a best effort and weird genomic phenomena could lead these
-// to not be 100% accurate
-function SequenceFeatureDetails(props: BaseProps) {
-  const { model, feature } = props
-  const { assemblyManager, rpcManager } = getSession(model)
-  const { assemblyNames } = model.view
-  const subfeatures = feature.subfeatures as Feat[]
-  const hasCDS = subfeatures.find(sub => sub.type === 'CDS')
-
-  const [sequence, setSequence] = useState<{
-    seq: string
-    upstream: string
-    downstream: string
-  }>()
-  const [error, setError] = useState<string>()
-  const [mode, setMode] = useState(hasCDS ? 'cds' : 'cdna')
-  const [assemblyName] = assemblyNames
-  const loading = !sequence
-
-  useEffect(() => {
-    async function fetchSeq(start: number, end: number, refName: string) {
-      const assembly = await assemblyManager.waitForAssembly(assemblyName)
-      if (!assembly) {
-        throw new Error('assembly not found')
-      }
-      const adapterConfig = getConf(assembly, ['sequence', 'adapter'])
-      const sessionId = 'getSequence'
-      const feats = await rpcManager.call(sessionId, 'CoreGetFeatures', {
-        adapterConfig,
-        sessionId,
-        region: { start, end, refName: assembly?.getCanonicalRefName(refName) },
-      })
-      const [feat] = feats as Feature[]
-      if (!feat) {
-        throw new Error('sequence not found')
-      }
-      return feat.get('seq') as string
-    }
-    ;(async () => {
-      try {
-        const { start, end, refName: ref } = feature
-        const seq = await fetchSeq(start, end, ref)
-        const upstream = await fetchSeq(Math.max(0, start - 500), start, ref)
-        const downstream = await fetchSeq(end, end + 500, ref)
-        setSequence({ seq, upstream, downstream })
-      } catch (e) {
-        setError(e)
-      }
-    })()
-  }, [feature, assemblyManager, rpcManager, assemblyName])
-
-  return (
-    <div>
-      <Select
-        value={mode}
-        onChange={event => setMode(event.target.value as string)}
-      >
-        {hasCDS ? <MenuItem value="cds">CDS</MenuItem> : null}
-        {hasCDS ? <MenuItem value="protein">Protein</MenuItem> : null}
-        <MenuItem value="gene">Gene with introns</MenuItem>
-        <MenuItem value="gene_collapsed_intron">
-          Gene w/ +/- 10bp of intron
-        </MenuItem>
-        <MenuItem value="gene_updownstream">Gene +/- 500bp upstream</MenuItem>
-        <MenuItem value="cdna">cDNA</MenuItem>
-      </Select>
-      <div style={{ display: 'inline' }}>
-        {error ? (
-          <Typography color="error">{error}</Typography>
-        ) : loading ? (
-          <div>Loading gene sequence...</div>
-        ) : sequence ? (
-          <div style={{ fontFamily: 'monospace', wordWrap: 'break-word' }}>
-            <SequencePanel feature={feature} mode={mode} sequence={sequence} />
-          </div>
-        ) : (
-          <div>No sequence found</div>
-        )}
-      </div>
     </div>
   )
 }
@@ -803,15 +418,15 @@ export const BaseFeatureDetails = observer((props: BaseInputProps) => {
 })
 
 export const FeatureDetails = (props: {
-  model: any
-  feature: any
+  model: IAnyStateTreeNode
+  feature: SimpleFeatureSerialized & { name?: string; id?: string }
   depth?: number
   omit?: string[]
   formatter?: (val: unknown, key: string) => JSX.Element
 }) => {
   const { model, feature, depth = 0 } = props
   const { name, id, type, subfeatures } = feature
-  const displayName = name || id
+  const displayName = (name || id) as string | undefined
   const ellipsedDisplayName =
     displayName && displayName.length > 20 ? '' : displayName
   const session = getSession(model)
